test: 사용자 상세정보 갱신 테스트

+++ app-main/src/main/java/net/causw/app/main/service/userInfo/UserInfoService.java
@@ -111,7 +111,10 @@ public class UserInfoService {
     List<String> deletingIdList = userCareerRepository.findAllCareerByUserInfoId(userInfo.getId()).stream()
         .map(BaseEntity::getId)
         .filter(id -> !requestedIdSet.contains(id)).toList();
-    userCareerRepository.deleteAllByIdInBatch(deletingIdList);
+
+    if (!deletingIdList.isEmpty()) {
+      userCareerRepository.deleteAllByIdInBatch(deletingIdList);
+    }
 
     return UserDtoMapper.INSTANCE.toUserInfoResponseDto(userInfo);
   }
